JUS-83 Fixed a minor issue with upload image

+++ server/controllers/eventController.js
@@ -21,6 +21,9 @@ const getEventImage = async (req, res, next) => {
 
 const addEvent = async (req, res, next) => {
     const errors = validationResult(req);
+    if (!req.file) {
+        return clientError(res, 'Invalid file provided.');
+    }
     if (!errors.isEmpty()) {
         eventImageService.deleteImage(req.file.path);
         return clientError(res, errors.array());
@@ -96,7 +99,7 @@ const getOrganizerEvents =  async (req, res, next) => {
                     count: output.length,
                     events: output.map(out => {
                         return {
-                            id: out._id, // output
+                            id: out._id, 
                             name: out.eventInfo.name,
                             description: out.eventInfo.description,
                             address:{
@@ -126,14 +129,11 @@ const updateEvents = async (req, res, next) => {
         return clientError(res, errors.array());
     }
     Event.findOne({ _id: req.query.eventId}, (err, event) => {
-    // Event.findOne({ _id: req.body.eventId}, (err, event) => {
         if (!event) {return clientError(res, "No such event exists ");}
         else if (event.eventInfo.status == "Ongoing" || event.eventInfo.status == "Completed"){
           return clientError(res, "event cannot be updated");
         }
-        // let deletepath = event.eventImage_path
         let deletepath = event.eventImagePath
-        console.log(deletepath)
         if (req.file){
             const update = {
                 eventInfo: {
